Fixed dtype of random_color and random_grayscale outputs

+++ imgproc/random.py
@@ -132,8 +132,8 @@ def random_hex(size):
 
 
 def random_color():
-	return np.random.randint(256, size=(3,))
+	return np.random.randint(256, size=(3,)).astype(np.uint8)
 
 
 def random_grayscale():
-	return np.array([random.randint(0, 255)] * 3)
+	return np.array([random.randint(0, 255)] * 3).astype(np.uint8)
